requirements.txt
bs4
requests
urllib3
pyyaml
html5lib
PyPDF2
pdfkit